resnet/requirements.txt
fire==0.5.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
opencv-python==4.11.0.86
matplotlib==3.10.1
openpyxl==3.1.5
Pillow==10.4.0
tensorboard==2.19.0
tqdm==4.67.1
